Изменения 1. Добавил RootFlow -builder -RootTabBarController 2. Удалил ViewController 3. Прописал основным контролером RootFlow в SceneDelegate
