+++ FitNote/Flows/RootFlow/View/RootTabBarController.swift
@@ -1,0 +1,16 @@
+//
+//  RootTabBarController.swift
+//  FitNote
+//
+//  Created by Роман Вертячих on 24.04.2025.
+//
+
+import UIKit
+
+final class RootTabBarController: UITabBarController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        tabBar.backgroundColor = .backgroundPrimary
+    }
+}
