--- FitNote/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  FitNote
-//
-//  Created by Роман Вертячих on 12.04.2025.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
